// src/app/shared/models/User.ts
export class User {
   
    constructor() {
       this.ActiveToDate =""        
         
    }
    UserID: number;
    Username: string;
    UserFullName: string;
    Phone: string;
    Email: string;
    Password: string;
    IsActive: boolean;
    ActiveToDate: string;
    PerGroupID: number;
    DepartmentID: number;
    Mobile: string;
    PhoneSwitch: string;
    PayrollEmpID: number;
    TajDbUserId: string;
    VacSigEmp: boolean;
    SigPassword: string;
    RankID: number;
    IsDeleted: boolean;
    VacNo: number;
    UserImage :any;
    UserImageMarking :any;
    UserImageSignature :any;
    UserImageNoSignature :any;
    UserImageURL :string;
    UserImageMarkingURL :string;
    UserImageNoSignatureURL :string;
    UserImageSignatureURL :string;
    IsAgent:boolean;
} 